pyproject.toml
[tool.poetry]
name = "dl4bi"
version = "0.1.0"
description = "Deep Learning for Bayesian Inference (DL4BI)"
authors = ["Daniel Jenson <daniel.a.jenson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
arviz = "^0.20.0"
datasets = "^3.2.0"
einops = "^0.8.0"
flax = "^0.10.2"
hydra-core = "^1.3.2"
imageio = "^2.35.1"
jraph = "0.0.6.dev0"
matplotlib = "^3.9.3"
numpy = "^2.0.0"
optax = "^0.2.4"
orbax = "^0.1.9"
orbax-checkpoint = "^0.11.0"
pandas = "^2.2.3"
pillow = "^11.0.0"
pygifsicle = "^1.1.0"
python = "^3.11"
scikit-learn = "^1.6.1"
scipy = "^1.14.1"
scoringrules = "^0.7.1"
sps = {git = "ssh://git@github.com/MLGlobalHealth/sps.git"}
tensorflow = "^2.18.0"
tensorflow-datasets = "^4.9.7"
tiktoken = "^0.8.0"
tqdm = "^4.67.1"
ucimlrepo = "^0.0.7"
wandb = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.ruff.lint]
extend-select = ["I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
